backend/requirements.txt
fastapi==0.110.0
uvicorn==0.24.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.0
Pillow==10.2.0
pytesseract==0.3.10
pydantic==2.6.1
google-cloud-aiplatform==1.66.0
vertexai==1.66.0
google-generativeai==0.7.2
httpx==0.27.0
python-dotenv==1.0.1

